Remove interactive shell and unblock signals.
On Ubuntu, the default shell for /bin/sh is dash - this creates another
shell w/ monitor mode enabled (it looks like). This causes some issues
with processes that are spawned and cannot be sent signals.

Remove SHELL_CMD and just call execve() instead. Also, we need to
unblock signals before execve() so they can be sent to the child
processes.